+++ examples/Dependency/FirstDependency.php
@@ -1,0 +1,13 @@
+<?php
+
+
+namespace PlugRoute\Example\Dependency;
+
+
+class FirstDependency
+{
+    public function apresentation()
+    {
+        return "First dependency<br>";
+    }
+}