+++ examples/dependencies.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+    'PlugRoute\Example\Dependency\ConstructDependency' => new \PlugRoute\Example\Dependency\ConstructDependency(
+        new \PlugRoute\Example\Dependency\FirstDependency()
+    )
+];